Move checks into check.py
