+++ hopper/__init__.py
@@ -1,1 +1,3 @@
 from .hopper import run_cmds, logger
+
+from . import check  # This registers Check classes
